requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
asgiref==3.8.111
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.34.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
CacheControl==0.14.2
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.3.0
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
duplicity==0.8.21
exceptiongroup==1.2.2
fastapi==0.115.12
fasteners==0.14.1
filelock==3.18.0
firebase-admin==6.7.0
frozenlist==1.6.0
fsspec==2025.3.2
future==0.18.2
gcloud==0.18.3
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
google-search-results==2.4.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hf-xet==1.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.31.1
hyperframe==6.1.0
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
keyring==23.5.0
langchain-core==0.3.54
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.32
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mongoengine==0.29.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
netifaces==0.11.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openai==1.78.1
opencv-python==4.11.0.86
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
portalocker==2.10.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycryptodome==3.22.0
pycups==2.0.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.9.0
pymacaroons==0.13.0
pymongo==3.11.4
PyNaCl==1.5.0
pyparsing==2.4.7
Pyrebase4==4.8.0
pyRFC3339==1.1
pytesseract==0.3.13
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.1.0
python-jwt==4.1.0
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qdrant-client==1.14.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
SecretStorage==3.3.1
sentence-transformers==4.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.5.6
social-auth-storage-mongoengine==1.0.1
speedtest-cli==2.1.3
sqlparse==0.5.3
starlette==0.46.2
stripe==12.1.0
sympy==1.14.0
systemd-python==234
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.4.0
usb-creator==0.3.7
uvicorn==0.34.2
wadllib==1.3.6
whisper==1.1.10
whitenoise==6.9.0
xdg==5
xkit==0.0.0
xxhash==3.5.0
yarl==1.20.0
zipp==1.0.0
zstandard==0.23.0
